+++ src/db/auth/auth.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
-import { AuthService } from './auth.service';
+import { AuthDbService } from './auth.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User])],
-
-  providers: [AuthService],
+  exports: [AuthDbService],
+  providers: [AuthDbService],
 })
-export class AuthModule {}
+export class AuthDbModule {}