+++ src/db/auth/auth.service.spec.ts
@@ -1,15 +1,15 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { AuthService } from './auth.service';
+import { AuthDbService } from './auth.service';
 
 describe('AuthService', () => {
-  let service: AuthService;
+  let service: AuthDbService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      providers: [AuthService],
+      providers: [AuthDbService],
     }).compile();
 
-    service = module.get<AuthService>(AuthService);
+    service = module.get<AuthDbService>(AuthDbService);
   });
 
   it('should be defined', () => {