+++ src/shared/errors/custom-exceptions.ts
@@ -1,5 +1,19 @@
 import { HttpException, HttpStatus } from '@nestjs/common';
 
+class BaseException extends HttpException {
+  constructor(
+    code: string,
+    error: any,
+    httpStatus: HttpStatus,
+    remainingTime?: number,
+    additionalValues?: Map<string, string>,
+  ) {
+    super({ code, error, remainingTime, additionalValues }, httpStatus);
+  }
+}
+
+export class CustomException extends BaseException {}
+
 export class CustomBadRequestException extends HttpException {
   constructor(code: string, message: string) {
     super({ code, message }, HttpStatus.BAD_REQUEST);