+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - Add basic modules, controllers and services
 - Add custom errors and changelog
 - Add connection with postgres and basic configuration and migrations
+- Add create user in db functionality